Move Stelline from tests to src/stelline.
