+++ Cargo.toml
@@ -40,6 +40,7 @@ smallvec       = { version = "1.3", optional = true }
 tokio          = { version = "1.33", optional = true, features = ["io-util", "macros", "net", "time", "sync", "rt-multi-thread" ] }
 tokio-rustls   = { version = "0.26", optional = true, default-features = false }
 tracing        = { version = "0.1.40", optional = true }
+tracing-subscriber = { version = "0.3.18", optional = true, features = ["env-filter"] }
 
 # For testing in integration tests:
 #mock_instant = { version = "0.4.0", optional = true }
@@ -62,6 +63,7 @@ zonefile    = ["bytes", "serde", "std"]
 # Unstable features
 unstable-client-transport = [ "moka", "net", "tracing" ]
 unstable-server-transport = ["arc-swap", "chrono/clock", "hex", "libc", "net", "tracing"]
+unstable-stelline = ["tracing", "tracing-subscriber", "unstable-server-transport", "zonefile"]
 unstable-zonetree = ["futures", "parking_lot", "serde", "tokio", "tracing"]
 
 # Test features
@@ -84,7 +86,6 @@ tokio              = { version = "1.37", features = ["rt-multi-thread", "io-util
 tokio-rustls       = { version = "0.26", default-features = false, features = [ "ring", "logging", "tls12" ] }
 tokio-test         = "0.4"
 tokio-tfo          = { version = "0.2.0" }
-tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 webpki-roots       = { version = "0.26" }
 
 # For the "mysql-zone" example
