+++ src/lib.rs
@@ -160,6 +160,7 @@ pub mod net;
 pub mod rdata;
 pub mod resolv;
 pub mod sign;
+pub mod stelline;
 pub mod test;
 pub mod tsig;
 pub mod utils;
