+++ src/stelline/client.rs
@@ -6,6 +6,8 @@ use std::pin::Pin;
 use std::rc::Rc;
 use std::sync::Mutex;
 use std::time::Duration;
+use std::vec::Vec;
+use std::boxed::Box;
 
 use bytes::Bytes;
 /*
@@ -15,16 +17,16 @@ use mock_instant::MockClock;
 use tracing::{debug, info_span, trace};
 use tracing_subscriber::EnvFilter;
 
-use domain::base::iana::Opcode;
-use domain::base::opt::{ComposeOptData, OptData};
-use domain::base::{Message, MessageBuilder};
-use domain::net::client::request::{
+use crate::base::iana::{Opcode, OptionCode};
+use crate::base::opt::{ComposeOptData, OptData};
+use crate::base::{Message, MessageBuilder};
+use crate::net::client::request::{
     ComposeRequest, Error, RequestMessage, SendRequest,
 };
 
-use crate::net::stelline::matches::match_msg;
-use crate::net::stelline::parse_query;
-use crate::net::stelline::parse_stelline::{
+use super::matches::match_msg;
+use super::parse_query;
+use super::parse_stelline::{
     Entry, Reply, Stelline, StepType,
 };
 
@@ -66,7 +68,7 @@ impl<'a> StellineError<'a> {
 
 #[derive(Debug)]
 pub enum StellineErrorCause {
-    ClientError(domain::net::client::request::Error),
+    ClientError(Error),
     MismatchedAnswer,
     MissingResponse,
     MissingStepEntry,
@@ -75,8 +77,8 @@ pub enum StellineErrorCause {
     MissingClient,
 }
 
-impl From<domain::net::client::request::Error> for StellineErrorCause {
-    fn from(err: domain::net::client::request::Error) -> Self {
+impl From<Error> for StellineErrorCause {
+    fn from(err: Error) -> Self {
         Self::ClientError(err)
     }
 }
@@ -567,7 +569,7 @@ struct RawOptData<'a> {
 }
 
 impl<'a> OptData for RawOptData<'a> {
-    fn code(&self) -> domain::base::iana::OptionCode {
+    fn code(&self) -> OptionCode {
         u16::from_be_bytes(self.bytes[0..2].try_into().unwrap()).into()
     }
 }
