+++ tests/net-client.rs
@@ -1,12 +1,11 @@
 #![cfg(feature = "net")]
-mod net;
 
-use crate::net::stelline::client::do_client_simple;
-use crate::net::stelline::client::CurrStepValue;
-use crate::net::stelline::connect::Connect;
-use crate::net::stelline::connection::Connection;
-use crate::net::stelline::dgram::Dgram;
-use crate::net::stelline::parse_stelline::parse_file;
+use domain::stelline::client::do_client_simple;
+use domain::stelline::client::CurrStepValue;
+use domain::stelline::connect::Connect;
+use domain::stelline::connection::Connection;
+use domain::stelline::dgram::Dgram;
+use domain::stelline::parse_stelline::parse_file;
 // use domain::net::client::clock::{Clock, FakeClock};
 use domain::net::client::dgram;
 use domain::net::client::dgram_stream;
